Add F# script to create a list of cultures
The list is limited to just a couple cultures, but can be easily expanded by adding more cultures or selecting all cultures.

The unity CultureInfos are quite outdated. For example Russia still
uses the Currency Symbol R instead of ₽.
Additional there are quite few cultures (113 in Unity vs. 832 in
.NET 4.6 on Windows 10).
And last but not least the data about the cultures in Unity is quite
limited: NumberFormat are often missing (for neutral cultures).